Enzo src-directory reorganization: following some of James's suggestions part 2 -> moved some files into a subdirectory called obsolete (maybe we can drop these in the future - but, that's a topic for a different PR)
