+++ src/Enzo/_enzo.hpp
@@ -195,7 +195,7 @@ extern "C" {
 #include "initial/EnzoInitialImplosion2.hpp"
 #include "initial/EnzoInitialInclinedWave.hpp"
 #include "io/EnzoInitialMusic.hpp"
-#include "initial/EnzoInitialPm.hpp"
+#include "obsolete/EnzoInitialPm.hpp"
 #include "initial/EnzoInitialPpmlTest.hpp"
 #include "initial/EnzoInitialSedovArray2.hpp"
 #include "initial/EnzoInitialSedovArray3.hpp"
@@ -260,7 +260,7 @@ extern "C" {
 #include "assorted/EnzoMethodGrackle.hpp"
 #include "gravity/EnzoMethodGravity.hpp"
 #include "assorted/EnzoMethodHeat.hpp"
-#include "hydro-mhd/EnzoMethodHydro.hpp"
+#include "obsolete/EnzoMethodHydro.hpp"
 #include "particle/formation/EnzoMethodMergeSinks.hpp"
 #include "hydro-mhd/EnzoMethodMHDVlct.hpp"
 #include "assorted/EnzoMethodM1Closure.hpp"
